+++ packages/BoMultLang/index.js
@@ -1,0 +1,7 @@
+import BoMultLang from './src/main';
+
+BoMultLang.install = function(Vue) {
+  Vue.component(BoMultLang.name, BoMultLang)
+};
+
+export default BoMultLang;